api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
pydantic==2.9.2
httpx==0.27.2
trafilatura==1.12.2
apscheduler==3.10.4
python-dotenv==1.0.1
